requirements.txt
adbc_driver_manager==1.1.0
adbc_driver_postgresql==1.1.0
adbc_driver_sqlite==1.1.0
AppKit==0.2.8
argcomplete==3.5.0
asgiref==3.8.1
atomicwrites==1.4.1
beautifulsoup4==4.12.3
boto3==1.35.0
botocore==1.35.0
Brotli==1.1.0
brotlicffi==1.1.0.0
celery==5.4.0
chardet==5.2.0
checks==0.2
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.8
ConfigParser==7.0.0
cryptography==41.0.7
cx_Oracle==8.3.0
cycler==0.12.1
Cython==3.0.11
dash_leaflet==1.0.15
dash_testing_stub==0.0.2
dask==2024.8.1
diskcache==5.6.3
dl==0.1.0
docutils==0.21.2
eventlet==0.36.1
fast_dash==0.2.9
fastparquet==2024.5.0
fire==0.6.0
flask_compress==1.15
Foundation==0.1.0a0.dev1
fsspec==2024.6.1
gevent==24.2.1
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.111.1
inotify==0.2.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jnius==1.1.0
js==1.0
keyring==25.3.0
lxml==5.1.0
matplotlib==3.8.4
mtrand==0.1
multiprocess==0.70.16
numexpr==2.8.5
objgraph==3.6.1
odfpy==1.4.1
openpyxl==3.1.5
ordereddict==1.1
orjson==3.10.7
PasteScript==3.6.0
percy==2.0.2
pexpect==4.9.0
pickle5==0.0.12
Pillow==10.4.0
protobuf==5.27.3
psutil==5.9.8
PyInstaller==6.10.0
pyodide==0.0.2
pyOpenSSL==24.2.1
PyQt4==4.11.4
PyQt5==5.15.11
pysqlcipher3==1.2.0
python-dotenv==1.0.1
python_calamine==0.2.3
pyxlsb==1.0.10
qtpy==2.4.1
railroad==0.5.0
redis==5.0.8
ruamel.yaml.clib==0.2.7
s3fs==2024.6.1
scikit_learn==1.5.1
scipy==1.14.0
selenium==4.23.1
setproctitle==1.3.3
sets==0.3.2
simplejson==3.19.3
skimage==0.0
sniffio==1.3.1
Sphinx==8.0.2
sqlcipher3==0.5.2
statsmodels==0.14.2
tables==3.9.2
thread==2.0.5
threadpoolctl==3.5.0
tornado==6.3.3
traitlets==5.14.3
trio==0.26.2
trove_classifiers==2024.7.2
Twisted==24.7.0
ukkonen==1.0.1
urllib3_secure_extra==0.1.0
vincent==0.4.4
watchdog==4.0.2
xarray==2024.7.0
xlrd==2.0.1
xlsxwriter==3.2.0
xmlrpclib==1.0.1
zope==5.10
zstandard==0.22.0
gunicorn
pandas
dash-tools
